feat: implement with serverless-webpack plagin
