+++ package.json
@@ -3,8 +3,7 @@
   "version": "1.0.0",
   "description": "",
   "scripts": {
-    "build": "webpack --config webpack.config.js",
-    "deploy": "npm run build && sls deploy",
+    "deploy": "sls deploy",
     "remove": "sls remove",
     "local": "sls offline start",
     "test": "jest",
@@ -18,6 +17,7 @@
     "@babel/preset-env": "^7.24.4",
     "@stylistic/eslint-plugin-js": "^1.7.2",
     "@testing-library/jest-dom": "^5.16.4",
+    "@types/node": "^20.12.7",
     "@typescript-eslint/eslint-plugin": "^7.7.0",
     "@typescript-eslint/parser": "^7.7.0",
     "aws-sdk-mock": "^5.9.0",
@@ -32,6 +32,7 @@
     "serverless-finch": "^4.0.0",
     "serverless-offline": "^13.3.4",
     "serverless-s3-cleaner": "^2.0.1",
+    "serverless-webpack": "^5.13.0",
     "ts-jest": "^29.1.2",
     "ts-loader": "^9.5.1",
     "ts-node": "^10.9.2",
