--- src/product-service/handlers/getProductsById.ts
@@ -1,36 +1,0 @@
-import products from "../data.json";
-import { APIGatewayProxyEvent } from "aws-lambda";
-
-export const getProductsById = async (event: APIGatewayProxyEvent) => {
-
-  const productId = event.pathParameters?.productId;
-
-  if (!productId) {
-
-    return {
-      statusCode: 400,
-      body: JSON.stringify({ message: "ProductId dosnt exist" }),
-    };
-  }
-
-  const product = products.find((product) => product.id === productId);
-
-  if (!product) {
-
-    return {
-      statusCode: 404,
-      headers: {
-        "Access-Control-Allow-Origin": "*",
-      },
-      body: "Product not found",
-    };
-  }
-
-  return {
-    statusCode: 200,
-    headers: {
-      "Access-Control-Allow-Origin": "*",
-    },
-    body: JSON.stringify(product),
-  };
-};
