+++ tsconfig.json
@@ -1,15 +1,12 @@
 {
-    "compilerOptions": {
-      "module": "commonjs",
-      "outDir": "dist",
-      "target": "es2017",
-      "noImplicitAny": true,
-      "strict": true,
-      "moduleResolution": "node",
-      "esModuleInterop": true,
-      "resolveJsonModule": true,
-      "sourceMap": true
-    },
-    "include": ["src/**/*"],
-    "exclude": ["node_modules"]
-  }
+  "compilerOptions": {
+    "target": "ES2020", 
+    "module": "commonjs", 
+    "lib": ["ES2020"], 
+    "strict": true, 
+    "esModuleInterop": true, 
+    "skipLibCheck": true, 
+    "forceConsistentCasingInFileNames": true,
+    "resolveJsonModule": true,
+}
+}
